consolidated in `control` subdirectory: - controlAddedListener - setControlTitle
