+++ public/leaflet-fullscreen/control/set-control-title.js
