+++ public/leaflet-fullscreen/fullscreen-map.js
@@ -4,7 +4,7 @@ import {
   map as leafletMap,
 } from '../leaflet/leaflet-src.esm.js'
 
-import { controlAddedListener } from './control-added-listener.js'
+import { controlAddedListener } from './control/control-added-listener.js'
 import { joinClassNames } from './join-class-names.js'
 import { mapLifecycleListener } from './map-lifecycle-listener.js'
 import { useBoolean } from './state/use-boolean.js'
